filename changes
Due to changes in the Intel compiler options, it is no longer possible to have free-form fortran with a .f suffix.  The old way is to use the

-ffree-form

flag + the .f suffix.  This becomes obsolete with Intel compiler 2019 and beyond.  Now the filenames are changed with .f90 and the flags are simplified.